+++ tests/objects/test_insole_on_conveyor.py
@@ -1,7 +1,7 @@
 import numpy as np
 import pybullet as pb
 import pytest
-from deformable_gym.envs.pybullet_tools import BulletSimulation
+from deformable_gym.envs.bullet_simulation import BulletSimulation
 from deformable_gym.objects.bullet_object import ObjectFactory, UrdfObject
 from numpy.testing import assert_array_almost_equal
